Refactor Sam API provider
